+++ src/Game/Text/fr/ScreenText.fr.sjson.lua
@@ -31,7 +31,7 @@ local newData = {
 	},
 	-- #endregion
 
-	-- #region Codex/Inventory
+	-- #region Codex/Inventory/Cauldron
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_MarketScreen_Resources",
 	-- 	DisplayName = "Barter Exchange",