+++ src/Scripts/RequirementsData.lua
@@ -15,6 +15,21 @@ local newNamedRequirements = {
 		{
 			PathFalse = { "GameState", "TextLinesRecord", "PersephoneReturnsHome01" },
 		},
-	}
+	},
+	-- If the first modded incantation can be unlocked or not - also controls if the cauldron category shows up
+	ModsNikkelMHadesBiomes_UnlockInRunWellShopsIncantation_Unlockable = {
+		{
+			-- Player already has Wells of Charon in vanilla runs
+			PathTrue = { "GameState", "WorldUpgradesAdded", "WorldUpgradeWellShops" },
+		},
+		{
+			PathTrue = { "GameState", "RoomsEntered", "A_Boss01" },
+		},
+		{
+			Path = { "GameState", "ModsNikkelMHadesBiomesCompletedRunsCache" },
+			Comparison = ">=",
+			Value = 1,
+		},
+	},
 }
 mod.AddTableKeysSkipDupes(game.NamedRequirementsData, newNamedRequirements)