+++ src/main.lua
@@ -354,6 +354,7 @@ local function on_ready()
 			import "Scripts/DeathLoopData.lua"
 			import "Scripts/EffectData.lua"
 			import "Scripts/GardenData.lua"
+			import "Scripts/GhostAdminData_Items.lua"
 			import "Scripts/HarvestData.lua"
 			import "Scripts/HeroData.lua"
 			import "Scripts/KeywordData.lua"